add order api

+++ APIs/CusCake.Application/Utils/DeliveryFeeCalculator.cs
@@ -1,23 +1,24 @@
 
 namespace CusCake.Application.Utils;
+
 public static class DeliveryFeeCalculator
 {
     // Danh sách bảng giá động (có thể dễ dàng chỉnh sửa)
-    private static readonly List<Tuple<decimal, decimal>> FeeTable =
+    private static readonly List<Tuple<double, double>> FeeTable =
     [
-        Tuple.Create(3m, 30000m),
-        Tuple.Create(6m, 35000m),
-        Tuple.Create(7m, 40000m),
-        Tuple.Create(8m, 45000m),
-        Tuple.Create(9m, 50000m),
-        Tuple.Create(10m, 55000m),
-        Tuple.Create(12m, 65000m),
-        Tuple.Create(13m, 75000m)
+        Tuple.Create(3.0, 30000.0),
+        Tuple.Create(6.0, 35000.0),
+        Tuple.Create(7.0, 40000.0),
+        Tuple.Create(8.0, 45000.0),
+        Tuple.Create(9.0, 50000.0),
+        Tuple.Create(10.0, 55000.0),
+        Tuple.Create(12.0, 65000.0),
+        Tuple.Create(13.0, 75000.0)
     ];
 
-    private const decimal Over13KmRate = 6000m; // 6.000 VNĐ/km
+    private const double Over13KmRate = 6000.0; // 6.000 VNĐ/km
 
-    public static decimal CalculateFee(decimal distanceInKm)
+    public static double CalculateFee(double distanceInKm)
     {
         foreach (var (maxDistance, fee) in FeeTable)
         {
@@ -26,8 +27,8 @@ public static class DeliveryFeeCalculator
         }
 
         // Nếu trên 13km, tính theo công thức động
-        decimal baseFee = FeeTable[^1].Item2; // Lấy mức phí cuối cùng (75000)
-        decimal extraDistance = distanceInKm - 13;
+        double baseFee = FeeTable[^1].Item2; // Lấy mức phí cuối cùng (75000)
+        double extraDistance = distanceInKm - 13.0;
         return baseFee + extraDistance * Over13KmRate;
     }
 }
